+++ stan/math/prim/scal/fun/promote_scalar.hpp
@@ -3,10 +3,8 @@
 
 #include <stan/math/prim/scal/fun/promote_scalar_type.hpp>
 #include <stan/math/prim/scal/meta/index_type.hpp>
-#include <vector>
 
 namespace stan {
-
   namespace math {
 
     /**
@@ -56,36 +54,6 @@ namespace stan {
     };
 
     /**
-     * Struct to hold static function for promoting underlying scalar
-     * types.  This specialization is for standard vector inputs.
-     *
-     * @tparam T return scalar type
-     * @tparam S input type for standard vector elements in static
-     * nested function, which must have an underlying scalar type
-     * assignable to T.
-     */
-    template <typename T, typename S>
-    struct promote_scalar_struct<T, std::vector<S> > {
-      /**
-       * Return the standard vector consisting of the recursive
-       * promotion of the elements of the input standard vector to the
-       * scalar type specified by the return template parameter.
-       *
-       * @param x input standard vector.
-       * @return standard vector with values promoted from input vector.
-       */
-      static std::vector<typename promote_scalar_type<T, S>::type>
-      apply(const std::vector<S>& x) {
-        typedef std::vector<typename promote_scalar_type<T, S>::type> return_t;
-        typedef typename index_type<return_t>::type idx_t;
-        return_t y(x.size());
-        for (idx_t i = 0; i < x.size(); ++i)
-          y[i] = promote_scalar_struct<T, S>::apply(x[i]);
-        return y;
-      }
-    };
-
-    /**
      * This is the top-level function to call to promote the scalar
      * types of an input of type S to type T.
      *
@@ -100,8 +68,6 @@ namespace stan {
       return promote_scalar_struct<T, S>::apply(x);
     }
 
-
   }
 }
-
 #endif
