+++ stan/math/prim/scal/meta/scalar_type_pre.hpp
@@ -2,9 +2,6 @@
 #define STAN_MATH_PRIM_SCAL_META_SCALAR_TYPE_PRE_HPP
 
 #include <stan/math/prim/scal/meta/is_vector.hpp>
-#include <stan/math/prim/mat/meta/is_vector.hpp>
-#include <stan/math/prim/arr/meta/is_vector.hpp>
-#include <stan/math/prim/mat/meta/value_type.hpp>
 #include <stan/math/prim/scal/meta/value_type.hpp>
 
 namespace stan {
