+++ test/unit/math/prim/scal/fun/promote_scalar_type_test.cpp
@@ -11,12 +11,3 @@ TEST(MathFunctionsPromoteScalarType,primitive) {
   expect_promote_type<vector<double>,
                       double, vector<int> >();
 }
-
-TEST(MathFunctionsPromoteScalarType,StdVector) {
-  using std::vector;
-  expect_promote_type<vector<vector<double> >, 
-                    double, vector<vector<int> > >();
-  expect_promote_type<vector<vector<double> >, 
-                    double, vector<vector<double> > >();
-}
-
